bug fix for commit 94c9d8, which introduced an oscillation in the placeholder text.
The code would clear the placeholder test to empty string when trying to
set the selection to the same value it was already set to.

Normally this would be an un-noticeable bug, but on a complex render
with other timers this shows as a flicker as the text transitions from
the placeholder text (Select...) to the correct text.